+++ src/Sienar.Plugin.Cms.Client/Identity/Processors/ClientDeleteAccountProcessor.cs
@@ -5,6 +5,7 @@ using Sienar.Data;
 using Sienar.Identity.Requests;
 using Sienar.Infrastructure;
 using Sienar.Processors;
+using Sienar.Security;
 
 namespace Sienar.Identity.Processors;
