+++ src/Sienar.Plugin.Cms.Client/Identity/Processors/ClientLogoutProcessor.cs
@@ -6,6 +6,7 @@ using Sienar.Extensions;
 using Sienar.Identity.Requests;
 using Sienar.Infrastructure;
 using Sienar.Processors;
+using Sienar.Security;
 
 namespace Sienar.Identity.Processors;
