+++ src/Sienar.Utils/Security/AuthStateProvider.cs
@@ -3,7 +3,7 @@ using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components.Authorization;
 
-namespace Sienar.Infrastructure;
+namespace Sienar.Security;
 
 /// <summary>
 /// A WASM-friendly implementation of <see cref="AuthenticationStateProvider"/>
